Fixed the windows store project.

+++ Emgu.CV.Extern/photo/photo_c.cpp
@@ -55,7 +55,7 @@ void cveIlluminationChange(cv::_InputArray* src, cv::_InputArray* mask, cv::_Out
    cv::illuminationChange(*src, mask ? *mask : (cv::InputArray) cv::noArray(), *dst, alpha, beta);
 }
 
-void cveTextureFlattening(cv::_InputArray* src, cv::_InputArray* mask, cv::_OutputArray* dst, double lowThreshold, double highThreshold, int kernelSize)
+void cveTextureFlattening(cv::_InputArray* src, cv::_InputArray* mask, cv::_OutputArray* dst, float lowThreshold, float highThreshold, int kernelSize)
 {
    cv::textureFlattening(*src, mask ? *mask : (cv::InputArray) cv::noArray(), *dst, lowThreshold, highThreshold, kernelSize);
 }
